+++ src/imio/pm/wsclient/browser/vocabularies.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 
+from collective.behavior.talcondition.utils import _evaluateExpression
 from datetime import datetime
 from imio.pm.wsclient import WS4PMClientMessageFactory as _
 from imio.pm.wsclient.config import CAN_NOT_CREATE_FOR_PROPOSING_GROUP_ERROR
@@ -125,11 +126,11 @@ class proposing_groups_for_user_vocabulary(object):
             # try to find out if a proposingGroup is forced in the configuration
             if field_mapping[u"field_name"] == "proposingGroup":
                 try:
-                    forcedProposingGroup = ws4pmsettings.renderTALExpression(
-                        context, portal, field_mapping["expression"], vars
+                    forcedProposingGroup = _evaluateExpression(
+                        context, expression=field_mapping["expression"], extra_expr_ctx=vars
                     )
                     break
-                except Exception, e:
+                except Exception as e:
                     portal.REQUEST.set("error_in_vocabularies", True)
                     IStatusMessage(portal.REQUEST).addStatusMessage(
                         _(
@@ -210,11 +211,11 @@ class categories_for_user_vocabulary(object):
             # try to find out if a proposingGroup is forced in the configuration
             if field_mapping[u"field_name"] == "category":
                 try:
-                    forcedCategory = ws4pmsettings.renderTALExpression(
-                        context, portal, field_mapping["expression"], vars
+                    forcedCategory = _evaluateExpression(
+                        context, expression=field_mapping["expression"], extra_expr_ctx=vars
                     )
                     break
-                except Exception, e:
+                except Exception as e:
                     portal.REQUEST.set("error_in_vocabularies", True)
                     IStatusMessage(portal.REQUEST).addStatusMessage(
                         _(
@@ -309,21 +310,16 @@ class desired_meetingdates_vocabulary(object):
         local = pytz.timezone("Europe/Brussels")
         if not possible_meetings:
             return SimpleVocabulary([])
-        for meeting in possible_meetings:
-            meeting["date"] = datetime.strptime(meeting["date"], "%Y-%m-%dT%H:%M:%S")
-            meeting["date"] = local.localize(meeting["date"])
-            meeting['date'] = meeting['date'].astimezone(local)
         terms = []
         allowed_meetings = queryMultiAdapter((context, possible_meetings), IPreferredMeetings)
         meetings = allowed_meetings and allowed_meetings.get() or possible_meetings
         for meeting_info in meetings:
-            display_date = datetime.strftime(meeting_info["date"], "%d/%m/%Y %H:%M") \
-                if isinstance(meeting_info["date"], datetime) else meeting_info["date"]
             terms.append(
                 SimpleTerm(
                     unicode(meeting_info["UID"]),
                     unicode(meeting_info["UID"]),
-                    unicode(display_date),
+                    # unicode(meeting_info["title"]),
+                    unicode(meeting_info["formatted_date"]),
                 )
             )
         return SimpleVocabulary(terms)