+++ src/imio/pm/wsclient/tests/testSettings.py
@@ -150,29 +150,10 @@ class testSettings(unittest.TestCase):
                                    'pm_userid': u'pmCreator1'}, ]
         self.assertEquals(ws4pmSettings._getUserIdToUseInTheNameOfWith(), 'lambda')
 
-    def test_renderTALExpression(self):
-        """
-          Test the method that will render a TAL expression
-        """
-        setRoles(self.portal, TEST_USER_ID, ('Manager',))
-        login(self.portal, TEST_USER_NAME)
-        # create an element to use in the TAL expression...
-        document = createDocument(self.portal)
-        ws4pmSettings = getMultiAdapter((self.portal, self.request), name='ws4pmclient-settings')
-        expr = u'python: None'
-        # make sure None is never returned by the renderer as it breaks SOAP calls
-        self.assertTrue(ws4pmSettings.renderTALExpression(document, self.portal, expr, {}) == u'')
-        expr = u'object/Title'
-        self.assertTrue(ws4pmSettings.renderTALExpression(document, self.portal, expr, {}) == u'Document title')
-        expr = u'string:"My expr result"'
-        self.assertTrue(ws4pmSettings.renderTALExpression(document, self.portal, expr, {}) == u'"My expr result"')
-        # with a wrong expression, we raise
-        expr = 'u object/wrongMethodCall'
-        self.assertRaises(CompilerError, ws4pmSettings.renderTALExpression, document, self.portal, expr, {})
-
 
 def test_suite():
-    from unittest import TestSuite, makeSuite
+    from unittest import makeSuite
+    from unittest import TestSuite
     suite = TestSuite()
     suite.addTest(makeSuite(testSettings))
     return suite